refactor(core): split flat-tree-layout into engine modules
Split the monolithic flat-tree-layout.ts into one module per
pipeline phase plus a thin shim for backwards compatibility:

  flat-tree/
    README.md       # algorithm spec
    types.ts        # shared internal types
    constants.ts    # gap derivations from PORT_LABEL_OUTER_REACH
    parents.ts      # primary parent + cycle breaking
    blocks.ts       # block partition + external-emitter detection
    internal.ts     # 3 block-internal layout variants
    sort.ts         # sibling sort by source-port label
    outer.ts        # block tree assembly
    spine.ts        # same-subgraph spine alignment
    hulls.ts        # subgraph hull bbox computation
    index.ts        # public entry, orchestrates the pipeline

  flat-tree-layout.ts  # thin re-export shim

No behaviour change. Build + all existing tests pass.

Next slices (tracked separately): per-module unit tests, edge
case fixtures, snapshot regression test, performance bench.
